+++ server/index.js
@@ -18,8 +18,8 @@ const {
   messageAllPlayers,
 } = require('./message-players');
 
-// Central store that keeps state of the whole game.
-const dataStore = {
+// Central store that keeps state of the whole game, server-side.
+global.dataStore = {
   npc: [],
   players: {}
 };
@@ -42,13 +42,13 @@ io.on('connection', (socket) => {
     _name = name;
 
     // Check if that name isn't used by a player.
-    if (dataStore.players[name] !== undefined) {
+    if (global.dataStore.players[name] !== undefined) {
       io.to(socket.id).emit('nameNotAvailable', name);
       return;
     }
 
     // Create the new player.
-    dataStore.players[name] = new Player({ name });
+    global.dataStore.players[name] = new Player({ name });
 
     // Messages
     greetSinglePlayer(io, socket.id, name);
@@ -56,21 +56,21 @@ io.on('connection', (socket) => {
     messageAllPlayers(socket, io);
 
     // NPC.
-    spawnMultipleNpc(io, socket.id, dataStore.npc);
+    spawnMultipleNpc(io, socket.id, global.dataStore.npc);
 
     // Players.
-    spawnPlayer(io, socket.id, dataStore.players[name]);
+    spawnPlayer(io, socket.id, global.dataStore.players[name]);
     updatePlayerState(socket);
 
     // Confirm the player has been created.
     // For the current player, spawn all previously existing players in game.
     io.to(socket.id).emit('playerCreated', {
       name,
-      players: dataStore.players
+      players: global.dataStore.players
     });
 
     // Spawn the current player on all existing player clients.
-    socket.broadcast.emit('addOtherPlayer', dataStore.players[name]);
+    socket.broadcast.emit('addOtherPlayer', global.dataStore.players[name]);
   });
 
   socket.on('disconnect', () => {
@@ -79,7 +79,7 @@ io.on('connection', (socket) => {
     socket.broadcast.emit('chatMessage', `${_name} has left.`);
 
     // Delete player.
-    delete dataStore.players[_name];
+    delete global.dataStore.players[_name];
 
     // Remove player from all other clients when he stops playing.
     socket.broadcast.emit('removePlayer', _name);
@@ -93,4 +93,4 @@ http.listen(port, () => {
 });
 
 // Start the server-side game.
-game({ io, dataStore });
+game({ io });